+++ R/internal.R
@@ -12,9 +12,8 @@ add_parameters <- function(x, object) {
   merge(x, object)
 }
 
-as_conditional_tibble <- function(x) {
-  if(requireNamespace("tibble", quietly = TRUE))
-    x <- tibble::as_tibble(x)
+as_tibble <- function(x) {
+  class(x) <- c("tbl_df", "tbl", "data.frame")
   x
 }
 
@@ -52,4 +51,15 @@ sum_fish <- function(x) {
   x[] <- lapply(x, sum)
   x[[1]] <- x[[1]] / nrow(x)
   x[1,]
+}
+
+population_names <- function(x) {
+  length <- vapply(x, function(x) length(unique(x)), 1L)
+  x <- x[length > 1L]
+  names <- names(x)
+  x <- apply(x, 1, function(x) sub("[.]", "_", x))
+  if(identical(length(names), 1L))
+    return(paste0(names, "_", x))
+  x <- apply(x, 2, function(x, names) paste0(names, "_", x), names = names)
+  apply(x, 2, function(x) paste0(x, collapse = "_"))
 }